scripts/requirements.txt
anthropic>=0.80.0
google-genai>=1.0.0
Pillow>=10.0.0
python-slugify>=8.0.0
